@@ -2,7 +2,7 @@
 using LightQueryProfiler.Shared.Repositories.Interfaces;
 using System.Data;
 using System.Data.Common;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 
 namespace LightQueryProfiler.Shared.Repositories
 {
